Add support for optional resources
Some entries can be enriched with external links.
These can be websites, PDFs, or videos.
This PR adds support for an optional `resources` field.
For now we only plan to use it on the website as the
rendering on the `README.md` could get a little cluttered.
Nevertheless I think it's a good idea to keep all data
in the repo.