Add machine-readable JSON output
Render a map as a JSON file where the key is the entry name and the
value is the entry data, which makes searching for a tool's data easier
when using tools like jq.

We use a different, de-normalized data format instead of the catalog, which
keeps the information for each tool in a struct instead of grouping tools by
tags.

Note that this data is not meant to be edited directly.